Refactor configuration management
- Centralized configuration logic into a new `config` package.
- Replaced inline environment variable handling with reusable utility functions.
- Simplified color printing by moving logic to a new `utils` package.
